Modify Todo.tsx template when path is /group/:id/todo and delete GroupTodo.tsx in src/templates
